@@ -10,7 +10,7 @@ authors = [
 ]
 description = "Students' solution checker"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 classifiers = [
     "Development Status :: 4 - Beta",
     "Typing :: Typed",
@@ -18,7 +18,6 @@ classifiers = [
     "Topic :: Education",
     "Topic :: Education :: Testing",
     "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
@@ -32,8 +31,8 @@ dependencies = [
     "rich >=13.0.0,<14.0.0",
     "python-gitlab >=3.0.0,<4.0.0",
     "requests >=2.20.0,<3.0.0",
-    "unshare >=0.22,<0.30; sys_platform != 'darwin'",
     "pytz >=2022.0,<2023.4; python_version < '3.9'",
+    "urllib3 >=2.0.0,<3.0.0",
 ]
 dynamic = ["version"]
 
@@ -46,7 +45,8 @@ test = [
     "pytest >=6.0.0,<8.0.0",
     "pytest-cov >=4.0.0,<5.0.0",
     "pytest-mock >=3.0.0,<4.0.0",
-    "black ==23.11.0",
+    "requests-mock >=1.0.0,<2.0.0",
+    "black ==23.12.1",
     "mypy >=1.0.0",
     "flake8 >=4.0.0,<7.0.0",
     "isort >=5.0.0,<7.0.0",
@@ -151,7 +151,8 @@ exclude = [
     "venv",
 
     "tests",
-    ".tmp"
+    ".tmp",
+    "__init__.py",
 ]
 
 # Allow autofix for all enabled rules (when `--fix`) is provided.
@@ -168,3 +169,7 @@ lines_after_imports = 2
 combine_as_imports = true
 default_section = "THIRDPARTY"
 known_first_party = "checker,tests"
+
+[tool.black]
+line-length = 120
+target-version = ['py312']
