@@ -1,15 +1,14 @@
 from __future__ import annotations
 
-import copy
-from typing import Any, Type
+from typing import Type
 
 import pytest
 
 from checker.configs import PipelineStageConfig
 from checker.exceptions import BadConfig, PluginExecutionFailed
+from checker.pipeline import PipelineRunner
 from checker.plugins import PluginABC
 from checker.plugins.base import PluginOutput
-from checker.pipeline import PipelineRunner
 
 
 class _FailPlugin(PluginABC):
@@ -26,7 +25,13 @@ class _ScorePlugin(PluginABC):
         score: float = 0.5
 
     def _run(self, args: Args, *, verbose: bool = False) -> PluginOutput:
-        return PluginOutput(output=f"Score: {args.score:.2f}", percentage=args.score)
+        if verbose:
+            return PluginOutput(
+                output=f"Score: {args.score:.2f}\nSome secret verbose line",
+                percentage=args.score,
+            )
+        else:
+            return PluginOutput(output=f"Score: {args.score:.2f}", percentage=args.score)
 
 
 class _EchoPlugin(PluginABC):
@@ -36,10 +41,7 @@ class _EchoPlugin(PluginABC):
         message: str
 
     def _run(self, args: Args, *, verbose: bool = False) -> PluginOutput:
-        if verbose:
-            return PluginOutput(output=args.message)
-        else:
-            return PluginOutput(output="")
+        return PluginOutput(output=args.message)
 
 
 @pytest.fixture
@@ -63,7 +65,7 @@ def sample_correct_pipeline() -> list[PipelineStageConfig]:
             name="stage2 - score",
             run="score",
             args={"score": 0.5},
-            register_output="score",
+            register_output="score_stage",
         ),
         PipelineStageConfig(
             name="stage3 - ignore fail",
@@ -71,14 +73,25 @@ def sample_correct_pipeline() -> list[PipelineStageConfig]:
             fail="never",
         ),
         PipelineStageConfig(
-            name="stage4 - skip fail if true",
+            name="stage4 - skip fail if run_if=False",
             run="fail",
-            run_if=True,
+            run_if=False,
         ),
         PipelineStageConfig(
-            name="stage4 - skip fail if registered output",
+            name="stage5 - skip echo if run_if=False",
+            run="echo",
+            args={"message": "skipped message"},
+            run_if=False,
+        ),
+        PipelineStageConfig(
+            name="stage6 - skip fail if registered output",
             run="fail",
-            run_if="${{ outputs.score.percentage > 0.7 }}",
+            run_if="${{ outputs.score_stage.percentage > 0.7 }}",
+        ),
+        PipelineStageConfig(
+            name="stage7 - second echo",
+            run="echo",
+            args={"message": "second message"},
         ),
     ]
 
@@ -88,7 +101,7 @@ class TestSampleFixtures:
         plugin = sample_plugins["echo"]()
         plugin.validate({"message": "Hello"})
         result = plugin.run({"message": "Hello"}, verbose=True)
-        assert result.percentage is None
+        assert result.percentage == 1.0
         assert result.output == "Hello"
 
         plugin = sample_plugins["score"]()
@@ -104,7 +117,7 @@ class TestSampleFixtures:
 
 
 class TestPipelineRunnerValidation:
-    def test_correct_pipeline(
+    def test_correct_pipeline_validation(
         self,
         sample_correct_pipeline: list[PipelineStageConfig],
         sample_plugins: dict[str, Type[PluginABC]],
@@ -121,7 +134,7 @@ class TestPipelineRunnerValidation:
 
     def test_unknown_plugin(self, sample_plugins: dict[str, Type[PluginABC]]) -> None:
         with pytest.raises(BadConfig) as exc_info:
-            pipeline_runner = PipelineRunner(
+            _ = PipelineRunner(
                 pipeline=[
                     PipelineStageConfig(
                         name="stage1 - echo",
@@ -134,32 +147,174 @@ class TestPipelineRunnerValidation:
             )
         assert "Unknown plugin" in str(exc_info.value)
 
-    # def test_validate_placeholders(self, sample_correct_pipeline: list[PipelineStageConfig]) -> None:
-    #     pipeline_runner = PipelineRunner(
-    #         pipeline=sample_correct_pipeline,
-    #         plugins={},
-    #         verbose=False,
-    #     )
-    #     with pytest.raises(BadConfig) as exc_info:
-    #         pipeline_runner.validate({}, validate_placeholders=True)
-    #     assert "Unknown plugin" in str(exc_info.value)
-    #
-    # def test_unknown_placeholder(self, sample_correct_pipeline: list[PipelineStageConfig], sample_plugins: dict[str, Type[PluginABC]]) -> None:
-    #     pipeline_runner = PipelineRunner(
-    #         pipeline=sample_correct_pipeline,
-    #         plugins=sample_plugins,
-    #         verbose=False,
-    #     )
-    #     with pytest.raises(BadConfig) as exc_info:
-    #         pipeline_runner.validate({}, validate_placeholders=True)
-    #     assert "Unknown placeholder" in str(exc_info.value)
-    #
-    # def test_invalid_run_if(self, sample_correct_pipeline: list[PipelineStageConfig], sample_plugins: dict[str, Type[PluginABC]]) -> None:
-    #     pipeline_runner = PipelineRunner(
-    #         pipeline=sample_correct_pipeline,
-    #         plugins=sample_plugins,
-    #         verbose=False,
-    #     )
-    #     with pytest.raises(BadConfig) as exc_info:
-    #         pipeline_runner.validate({"score": 0.5}, validate_placeholders=True)
-    #
+    def test_validate_placeholders(self, sample_correct_pipeline: list[PipelineStageConfig]) -> None:
+        with pytest.raises(BadConfig) as exc_info:
+            _ = PipelineRunner(
+                pipeline=sample_correct_pipeline,
+                plugins={},
+                verbose=False,
+            )
+        assert "Unknown plugin" in str(exc_info.value)
+
+    def test_unknown_placeholder(
+        self,
+        sample_correct_pipeline: list[PipelineStageConfig],
+        sample_plugins: dict[str, Type[PluginABC]],
+    ) -> None:
+        pipeline_runner = PipelineRunner(
+            pipeline=sample_correct_pipeline,
+            plugins=sample_plugins,
+            verbose=False,
+        )
+        with pytest.raises(BadConfig):
+            pipeline_runner.validate({}, validate_placeholders=True)
+        # TODO: fix it, now throwing Validation Error
+        # assert "Unknown placeholder" in str(exc_info.value)
+
+    def test_invalid_run_if(
+        self,
+        sample_correct_pipeline: list[PipelineStageConfig],
+        sample_plugins: dict[str, Type[PluginABC]],
+    ) -> None:
+        pipeline_runner = PipelineRunner(
+            pipeline=sample_correct_pipeline,
+            plugins=sample_plugins,
+            verbose=False,
+        )
+        with pytest.raises(BadConfig):
+            pipeline_runner.validate({"score": 0.5}, validate_placeholders=True)
+
+    def test_invalid_register_output(
+        self,
+        sample_correct_pipeline: list[PipelineStageConfig],
+        sample_plugins: dict[str, Type[PluginABC]],
+    ) -> None:
+        sample_correct_pipeline[1].register_output = "unknown"
+        pipeline_runner = PipelineRunner(
+            pipeline=sample_correct_pipeline,
+            plugins=sample_plugins,
+            verbose=False,
+        )
+        with pytest.raises(BadConfig) as exc_info:
+            pipeline_runner.validate({"message": "some valid message"}, validate_placeholders=True)
+        assert "Invalid template" in str(exc_info.value)
+
+    def test_run_correct_pipeline_verbose(
+        self,
+        sample_correct_pipeline: list[PipelineStageConfig],
+        sample_plugins: dict[str, Type[PluginABC]],
+        capsys: pytest.CaptureFixture[str],
+    ) -> None:
+        pipeline_runner = PipelineRunner(
+            pipeline=sample_correct_pipeline,
+            plugins=sample_plugins,
+            verbose=True,
+        )
+        result = pipeline_runner.run({"message": "Hello"})
+        assert not result.failed
+        captured = capsys.readouterr().err
+        assert "Hello" in captured
+        assert "Score: 0.50" in captured
+        assert "second message" in captured
+        # in args print, so in the verbose output
+        # assert "skipped message" not in captured.out and "skipped message" not in captured.err
+        # verbose output
+        assert "Some secret verbose line" in captured
+        # stages names are printed
+        for stage_name in ["stage1", "stage2", "stage3", "stage4", "stage5", "stage6"]:
+            assert stage_name in captured
+
+    def test_run_correct_pipeline_not_verbose(
+        self,
+        sample_correct_pipeline: list[PipelineStageConfig],
+        sample_plugins: dict[str, Type[PluginABC]],
+        capsys: pytest.CaptureFixture[str],
+    ) -> None:
+        pipeline_runner = PipelineRunner(
+            pipeline=sample_correct_pipeline,
+            plugins=sample_plugins,
+            verbose=False,
+        )
+        result = pipeline_runner.run({"message": "Hello"})
+        assert not result.failed
+        captured = capsys.readouterr().err
+        assert "Hello" in captured
+        assert "Score: 0.50" in captured
+        assert "second message" in captured
+        # in args print, so not in the non-verbose output
+        assert "skipped message" not in captured
+        # verbose output
+        assert "Some secret verbose line" not in captured
+        # stages names are printed
+        for stage_name in ["stage1", "stage2", "stage3", "stage4", "stage5", "stage6"]:
+            assert stage_name in captured
+
+    def test_dry_run(
+        self,
+        sample_correct_pipeline: list[PipelineStageConfig],
+        sample_plugins: dict[str, Type[PluginABC]],
+        capsys: pytest.CaptureFixture[str],
+    ) -> None:
+        pipeline_runner = PipelineRunner(
+            pipeline=sample_correct_pipeline,
+            plugins=sample_plugins,
+            verbose=False,
+        )
+        result = pipeline_runner.run({"message": "Hello"}, dry_run=True)
+        assert not result.failed
+        captured = capsys.readouterr().err
+        # no "error!" msg as all is skipped
+        assert "error!" not in captured
+        assert "[output here]" in captured
+        assert "dry run!" in captured
+        # stages names are printed
+        for stage_name in ["stage1", "stage2", "stage3", "stage4", "stage5", "stage6"]:
+            assert stage_name in captured
+
+    def test_fail_fast(
+        self,
+        sample_correct_pipeline: list[PipelineStageConfig],
+        sample_plugins: dict[str, Type[PluginABC]],
+        capsys: pytest.CaptureFixture[str],
+    ) -> None:
+        sample_correct_pipeline[2].fail = PipelineStageConfig.FailType.FAST
+        pipeline_runner = PipelineRunner(
+            pipeline=sample_correct_pipeline,
+            plugins=sample_plugins,
+            verbose=False,
+        )
+        result = pipeline_runner.run({"message": "Hello"})
+        assert result.failed
+        captured = capsys.readouterr().err
+        # no "error!" msg as all is skipped
+        assert "error!" in captured
+        # fist echo works, second not
+        assert "Hello" in captured
+        assert "second message" not in captured
+        # stages names are printed
+        for stage_name in ["stage1", "stage2", "stage3", "stage4", "stage5", "stage6"]:
+            assert stage_name in captured
+
+    def test_fail_after_all(
+        self,
+        sample_correct_pipeline: list[PipelineStageConfig],
+        sample_plugins: dict[str, Type[PluginABC]],
+        capsys: pytest.CaptureFixture[str],
+    ) -> None:
+        sample_correct_pipeline[2].fail = PipelineStageConfig.FailType.AFTER_ALL
+        pipeline_runner = PipelineRunner(
+            pipeline=sample_correct_pipeline,
+            plugins=sample_plugins,
+            verbose=False,
+        )
+        result = pipeline_runner.run({"message": "Hello"})
+        assert result.failed
+        captured = capsys.readouterr().err
+        # no "error!" msg as all is skipped
+        assert "error!" in captured
+        # fist echo works, second not
+        assert "Hello" in captured
+        assert "second message" in captured
+        # stages names are printed
+        for stage_name in ["stage1", "stage2", "stage3", "stage4", "stage5", "stage6"]:
+            assert stage_name in captured
