@@ -1,13 +1,4 @@
 from .checker import CheckerTestingConfig  # noqa: F401
-from .checker import (
-    CheckerConfig,
-    CheckerExportConfig,
-    PipelineStageConfig,
-)
-from .deadlines import (
-    DeadlinesConfig,
-    DeadlinesGroupConfig,
-    DeadlinesSettingsConfig,
-    DeadlinesTaskConfig,
-)  # noqa: F401
+from .checker import CheckerConfig, CheckerExportConfig, CheckerStructureConfig, PipelineStageConfig  # noqa: F401
+from .deadlines import DeadlinesConfig, DeadlinesGroupConfig, DeadlinesSettingsConfig, DeadlinesTaskConfig  # noqa: F401
 from .task import TaskConfig  # noqa: F401
