@@ -3,13 +3,11 @@ from __future__ import annotations
 import sys
 from datetime import datetime, timedelta
 from enum import Enum
-from typing import cast, Union, Optional
+from typing import Optional, Union
+
 
 if sys.version_info < (3, 8):
-    from pytz import (
-        ZoneInfoNotFoundError as ZoneInfoNotFoundError,
-        timezone as ZoneInfo,
-    )
+    from pytz import ZoneInfoNotFoundError as ZoneInfoNotFoundError, timezone as ZoneInfo
 else:
     from zoneinfo import ZoneInfo, ZoneInfoNotFoundError
 
@@ -96,20 +94,14 @@ class DeadlinesGroupConfig(CustomBaseModel):
         if isinstance(self.end, timedelta) and self.end < timedelta():
             raise ValueError(f"end timedelta <{self.end}> should be positive")
         if isinstance(self.end, datetime) and self.end < self.start:
-            raise ValueError(
-                f"end datetime <{self.end}> should be after the start <{self.start}>"
-            )
+            raise ValueError(f"end datetime <{self.end}> should be after the start <{self.start}>")
 
         # check steps
-        last_step_date_or_delta = self.start
+        last_step_date_or_delta: datetime | timedelta = self.start
         for _, date_or_delta in self.steps.items():
-            step_date = (
-                self.start + cast(timedelta, date_or_delta)
-                if isinstance(date_or_delta, timedelta)
-                else date_or_delta
-            )
+            step_date = self.start + date_or_delta if isinstance(date_or_delta, timedelta) else date_or_delta
             last_step_date = (
-                self.start + cast(timedelta, last_step_date_or_delta)
+                self.start + last_step_date_or_delta
                 if isinstance(last_step_date_or_delta, timedelta)
                 else last_step_date_or_delta
             )
@@ -117,20 +109,19 @@ class DeadlinesGroupConfig(CustomBaseModel):
             if isinstance(date_or_delta, timedelta) and date_or_delta < timedelta():
                 raise ValueError(f"step timedelta <{date_or_delta}> should be positive")
             if isinstance(date_or_delta, datetime) and date_or_delta <= self.start:
-                raise ValueError(
-                    f"step datetime <{date_or_delta}> should be after the start {self.start}"
-                )
+                raise ValueError(f"step datetime <{date_or_delta}> should be after the start {self.start}")
 
             if step_date <= last_step_date:
                 raise ValueError(
-                    f"step datetime/timedelta <{date_or_delta}> should be after the last step <{last_step_date_or_delta}>"
+                    f"step datetime/timedelta <{date_or_delta}> "
+                    f"should be after the last step <{last_step_date_or_delta}>"
                 )
             last_step_date_or_delta = date_or_delta
 
         return self
 
 
-class DeadlinesConfig(CustomBaseModel, YamlLoaderMixin['DeadlinesConfig']):
+class DeadlinesConfig(CustomBaseModel, YamlLoaderMixin["DeadlinesConfig"]):
     """Deadlines configuration."""
 
     version: int
@@ -155,9 +146,7 @@ class DeadlinesConfig(CustomBaseModel, YamlLoaderMixin['DeadlinesConfig']):
         self,
         enabled: bool | None = None,
     ) -> list[DeadlinesTaskConfig]:
-        tasks = [
-            task for group in self.get_groups(enabled=enabled) for task in group.tasks
-        ]
+        tasks = [task for group in self.get_groups(enabled=enabled) for task in group.tasks]
 
         if enabled is not None:
             tasks = [task for task in tasks if task.enabled == enabled]
@@ -175,9 +164,7 @@ class DeadlinesConfig(CustomBaseModel, YamlLoaderMixin['DeadlinesConfig']):
 
     @field_validator("schedule")
     @classmethod
-    def check_group_names_unique(
-        cls, data: list[DeadlinesGroupConfig]
-    ) -> list[DeadlinesGroupConfig]:
+    def check_group_names_unique(cls, data: list[DeadlinesGroupConfig]) -> list[DeadlinesGroupConfig]:
         groups = [group.name for group in data]
         duplicates = [name for name in groups if groups.count(name) > 1]
         if duplicates:
@@ -186,9 +173,7 @@ class DeadlinesConfig(CustomBaseModel, YamlLoaderMixin['DeadlinesConfig']):
 
     @field_validator("schedule")
     @classmethod
-    def check_task_names_unique(
-        cls, data: list[DeadlinesGroupConfig]
-    ) -> list[DeadlinesGroupConfig]:
+    def check_task_names_unique(cls, data: list[DeadlinesGroupConfig]) -> list[DeadlinesGroupConfig]:
         tasks_names = [task.name for group in data for task in group.tasks]
         duplicates = [name for name in tasks_names if tasks_names.count(name) > 1]
         if duplicates:
