@@ -1,8 +1,7 @@
 from __future__ import annotations
 
-import re
 from pathlib import Path
-from typing import Any, Generic, TypeVar, Type, Protocol
+from typing import Any, Generic, TypeVar
 
 import pydantic
 import yaml
@@ -14,34 +13,28 @@ class CustomBaseModel(pydantic.BaseModel):
     model_config = pydantic.ConfigDict(extra="forbid", validate_default=True)
 
 
-# class BaseModelProtocol(Protocol):
-#     def to_yaml(self: pydantic.BaseModel, path: Path) -> None:
-#         ...
-#
-#     def get_json_schema(self: pydantic.BaseModel) -> dict[str, Any]:
-#         ...
-
-
-T = TypeVar('T', bound=pydantic.BaseModel)
+T = TypeVar("T", bound=pydantic.BaseModel)
 
 
 class YamlLoaderMixin(Generic[T]):
     @classmethod
-    def from_yaml(cls: type[T], path: Path) -> T:
+    def from_yaml(cls: type[T], path: Path) -> T:  # type: ignore[misc]
         try:
             with path.open() as f:
                 return cls(**yaml.safe_load(f))
         except FileNotFoundError:
             raise BadConfig(f"File {path} not found")
+        except TypeError as e:
+            raise BadConfig(f"Config YAML error:\n{e}")
         except yaml.YAMLError as e:
             raise BadConfig(f"Config YAML error:\n{e}")
         except pydantic.ValidationError as e:
             raise BadConfig(f"Config Validation error:\n{e}")
 
-    def to_yaml(self: T, path: Path) -> None:
+    def to_yaml(self: T, path: Path) -> None:  # type: ignore[misc]
         with path.open("w") as f:
             yaml.dump(self.model_dump(), f)
 
     @classmethod
-    def get_json_schema(cls: type[T]) -> dict[str, Any]:
+    def get_json_schema(cls: type[T]) -> dict[str, Any]:  # type: ignore[misc]
         return cls.model_json_schema()
