@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import time
 from dataclasses import dataclass
-from typing import Any, Type
+from typing import Any
 
 import jinja2.nativetypes
 
@@ -14,15 +14,27 @@ from .utils import print_info
 
 @dataclass
 class PipelineStageResult:
+    """Result of a single pipeline stage.
+    :param name: name of the stage
+    :param failed: if True, stage failed
+    :param skipped: if True, stage was skipped
+    :param percentage: optional percentage of points earned
+    :param elapsed_time: optional elapsed time in seconds
+    :param output: output of the stage
+    """
+
     name: str
     failed: bool
     skipped: bool
-    percentage: float = 0.0
+    percentage: float | None = None
     elapsed_time: float | None = None
     output: str = ""
 
     def __str__(self) -> str:  # pragma: no cover
-        return f"PipelineStageResult: failed={int(self.failed)}, skipped={int(self.skipped)}, percentage={self.percentage:.2f}, name='{self.name}'"
+        return (
+            f"PipelineStageResult: failed={int(self.failed)}, "
+            f"skipped={int(self.skipped)}, percentage={self.percentage or 1.0:.2f}, name='{self.name}'"
+        )
 
 
 @dataclass
@@ -68,9 +80,7 @@ class ParametersResolver:
         elif isinstance(template, list):
             return [self.resolve(item, context) for item in template]
         elif isinstance(template, dict):
-            return {
-                key: self.resolve(value, context) for key, value in template.items()
-            }
+            return {key: self.resolve(value, context) for key, value in template.items()}
         else:
             return template
 
@@ -115,23 +125,17 @@ class PipelineRunner:
         for pipeline_stage in self.pipeline:
             # validate plugin exists
             if pipeline_stage.run not in self.plugins:
-                raise BadConfig(
-                    f"Unknown plugin {pipeline_stage.run} in pipeline stage {pipeline_stage.name}"
-                )
+                raise BadConfig(f"Unknown plugin {pipeline_stage.run} in pipeline stage {pipeline_stage.name}")
             plugin_class = self.plugins[pipeline_stage.run]
 
             # validate args of the plugin (first resolve placeholders)
             if validate_placeholders:
-                resolved_args = self.parameters_resolver.resolve(
-                    pipeline_stage.args, context
-                )
+                resolved_args = self.parameters_resolver.resolve(pipeline_stage.args, context)
                 plugin_class.validate(resolved_args)
 
             # validate run_if condition
             if validate_placeholders and pipeline_stage.run_if:
-                resolved_run_if = self.parameters_resolver.resolve(
-                    pipeline_stage.run_if, context
-                )
+                resolved_run_if = self.parameters_resolver.resolve(pipeline_stage.run_if, context)
                 if not isinstance(resolved_run_if, bool):
                     raise BadConfig(
                         f"Invalid run_if condition {pipeline_stage.run_if} in pipeline stage {pipeline_stage.name}"
@@ -139,12 +143,11 @@ class PipelineRunner:
 
             # add output to context if set register parameter
             if pipeline_stage.register_output:
-                context.setdefault("outputs", {})[
-                    pipeline_stage.register_output
-                ] = PipelineStageResult(
+                context.setdefault("outputs", {})[pipeline_stage.register_output] = PipelineStageResult(
                     name=pipeline_stage.name,
                     failed=False,
                     skipped=True,
+                    percentage=1.0,
                 )
 
     def run(
@@ -158,12 +161,10 @@ class PipelineRunner:
         skip_the_rest = False
         for pipeline_stage in self.pipeline:
             # resolve placeholders in arguments
-            resolved_args = self.parameters_resolver.resolve(
-                pipeline_stage.args, context=context
-            )
+            resolved_args = self.parameters_resolver.resolve(pipeline_stage.args, context=context)
             resolved_run_if = (
                 self.parameters_resolver.resolve(pipeline_stage.run_if, context=context)
-                if pipeline_stage.run_if
+                if pipeline_stage.run_if is not None
                 else None
             )
 
@@ -189,7 +190,7 @@ class PipelineRunner:
                 continue
 
             # resolve run condition if any; skip if run_if=False
-            if pipeline_stage.run_if:
+            if pipeline_stage.run_if is not None:
                 if not resolved_run_if:
                     print_info(f"skipped! (run_if={resolved_run_if})", color="blue")
                     pipeline_stage_results.append(
@@ -217,6 +218,11 @@ class PipelineRunner:
                         percentage=1.0,
                     )
                 )
+
+                # register output if required
+                if pipeline_stage.register_output:
+                    context.setdefault("outputs", {})[pipeline_stage.register_output] = pipeline_stage_results[-1]
+
                 continue
 
             # run the plugin with executor
@@ -225,9 +231,7 @@ class PipelineRunner:
                 result = plugin.run(resolved_args, verbose=self.verbose)
                 _end_time = time.perf_counter()
                 print_info(result.output or "[empty output]")
-                print_info(
-                    f"> elapsed time: {_end_time-_start_time:.2f}s", color="grey"
-                )
+                print_info(f"> elapsed time: {_end_time-_start_time:.2f}s", color="grey")
                 print_info("ok!", color="green")
                 pipeline_stage_results.append(
                     PipelineStageResult(
@@ -235,17 +239,14 @@ class PipelineRunner:
                         failed=False,
                         skipped=False,
                         output=result.output,
-                        percentage=1.0,  # TODO: get percentage from plugin
+                        percentage=result.percentage,
                         elapsed_time=_end_time - _start_time,
                     )
                 )
             except PluginExecutionFailed as e:
                 _end_time = time.perf_counter()
                 print_info(e.output or "[empty output]")
-                print_info(
-                    f"> elapsed time: {_end_time-_start_time:.2f}s", color="grey"
-                )
-                print_info("error!", color="red")
+                print_info(f"> elapsed time: {_end_time-_start_time:.2f}s", color="grey")
                 pipeline_stage_results.append(
                     PipelineStageResult(
                         name=pipeline_stage.name,
@@ -257,20 +258,21 @@ class PipelineRunner:
                     )
                 )
                 if pipeline_stage.fail == PipelineStageConfig.FailType.FAST:
+                    print_info("error! (now as fail=fast)", color="red")
                     skip_the_rest = True
                     pipeline_passed = False
                 elif pipeline_stage.fail == PipelineStageConfig.FailType.AFTER_ALL:
+                    print_info("error! (later as fail=after_all)", color="red")
                     pipeline_passed = False
                 elif pipeline_stage.fail == PipelineStageConfig.FailType.NEVER:
+                    print_info("error! (ignored as fail=never)", color="red")
                     pass
                 else:
-                    assert False, f"Unknown fail type {pipeline_stage.fail}"
+                    assert False, f"Unknown fail type {pipeline_stage.fail}"  # pragma: no cover
 
             # register output if required
             if pipeline_stage.register_output:
-                context.setdefault("outputs", {})[
-                    pipeline_stage.register_output
-                ] = pipeline_stage_results[-1]
+                context.setdefault("outputs", {})[pipeline_stage.register_output] = pipeline_stage_results[-1]
 
         return PipelineResult(
             failed=not pipeline_passed,
