@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from collections.abc import Callable
+from inspect import cleandoc
 from pathlib import Path
 from typing import Any
 
@@ -11,20 +12,23 @@ from checker.exceptions import PluginExecutionFailed
 from checker.plugins.regex import CheckRegexpsPlugin
 
 
+T_CREATE_TEST_FILES = Callable[[dict[str, str]], Path]
+
+
+@pytest.fixture
+def create_test_files(tmpdir: Path) -> T_CREATE_TEST_FILES:
+    def _create_test_files(files_content: dict[str, str]) -> Path:
+        for filename, content in files_content.items():
+            file = Path(tmpdir / filename)
+            file.parent.mkdir(parents=True, exist_ok=True)
+            with open(file, "w") as f:
+                f.write(cleandoc(content))
+        return tmpdir
+
+    return _create_test_files
+
+
 class TestCheckRegexpsPlugin:
-    T_CREATE_TEST_FILES = Callable[[dict[str, str]], Path]
-
-    @pytest.fixture
-    def create_test_files(self, tmpdir: Path) -> T_CREATE_TEST_FILES:
-        def _create_test_files(files_content: dict[str, str]) -> Path:
-            for filename, content in files_content.items():
-                file = tmpdir / filename
-                with open(file, "w") as f:
-                    f.write(content)
-            return tmpdir
-
-        return _create_test_files
-
     # TODO: add tests with wrong patterns and regexps
     @pytest.mark.parametrize(
         "parameters, expected_exception",
@@ -41,9 +45,7 @@ class TestCheckRegexpsPlugin:
             ),
         ],
     )
-    def test_plugin_args(
-        self, parameters: dict[str, Any], expected_exception: Exception | None
-    ) -> None:
+    def test_plugin_args(self, parameters: dict[str, Any], expected_exception: Exception | None) -> None:
         if expected_exception:
             with pytest.raises(expected_exception):
                 CheckRegexpsPlugin.Args(**parameters)
@@ -78,9 +80,7 @@ class TestCheckRegexpsPlugin:
         regexps = ["forbidden"]
 
         plugin = CheckRegexpsPlugin()
-        args = CheckRegexpsPlugin.Args(
-            origin=str(origin), patterns=patterns, regexps=regexps
-        )
+        args = CheckRegexpsPlugin.Args(origin=str(origin), patterns=patterns, regexps=regexps)
 
         if expected_exception:
             with pytest.raises(expected_exception):
@@ -115,9 +115,7 @@ class TestCheckRegexpsPlugin:
         patterns = ["*"]
 
         plugin = CheckRegexpsPlugin()
-        args = CheckRegexpsPlugin.Args(
-            origin=str(origin), patterns=patterns, regexps=regexps
-        )
+        args = CheckRegexpsPlugin.Args(origin=str(origin), patterns=patterns, regexps=regexps)
 
         if expected_exception:
             with pytest.raises(expected_exception) as exc_info:
@@ -125,18 +123,12 @@ class TestCheckRegexpsPlugin:
             assert "matches regexp" in str(exc_info.value)
         else:
             assert plugin._run(args).output == "No forbidden regexps found"
-            assert (
-                plugin._run(args, verbose=True).output == "No forbidden regexps found"
-            )
-            assert (
-                plugin._run(args, verbose=False).output == "No forbidden regexps found"
-            )
+            assert plugin._run(args, verbose=True).output == "No forbidden regexps found"
+            assert plugin._run(args, verbose=False).output == "No forbidden regexps found"
 
     def test_non_existent_origin(self) -> None:
         plugin = CheckRegexpsPlugin()
-        args = CheckRegexpsPlugin.Args(
-            origin="/tmp/non_existent", patterns=["*.txt"], regexps=["forbidden"]
-        )
+        args = CheckRegexpsPlugin.Args(origin="/tmp/non_existent", patterns=["*.txt"], regexps=["forbidden"])
 
         with pytest.raises(PluginExecutionFailed) as exc_info:
             plugin._run(args)
